# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AI IDS Project API",
        "description": "API documentation for AI IDS Project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/alerts/{id}": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Show alert details",
                "description": "Display a specific alert.",
                "operationId": "6b0af8639796a4977541ad512ae92da1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alert details returned"
                    },
                    "404": {
                        "description": "Alert not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Delete an alert",
                "description": "Delete an alert.",
                "operationId": "86cbb74b4dfcee5520381721a85833e6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Alert deleted successfully"
                    },
                    "404": {
                        "description": "Alert not found"
                    }
                }
            }
        },
...
    },
...
}